change type of transcation date to long in hierarchical transcation report

+++ Data/Data/DGINTFRAUD/ArtDgamlHierarchicalTransaction.cs
@@ -17,7 +17,7 @@ namespace Data.Data.DGINTFRAUD
         public string? Status { get; set; }
         public string? Transaction_Reference { get; set; }
         public string? Transaction_Type { get; set; }
-        public int? Transaction_Date { get; set; }
+        public long? Transaction_Date { get; set; }
         public decimal Base_Amount_EGP { get; set; }
         public decimal Equivalent_Amount { get; set; }
         public string? Equivalent_Currency { get; set; }
